java english iter_5

+++ src/main/java/org/636766f81a6d9265ec01775b/636766f81a6d9265ec01775b.java
@@ -13,13 +13,13 @@ public class CustomClassReader extends ClassReader {
      */
     public long readLong(final int offset) {
         byte[] b = this.b;
-        return (((long)b[offset] & 0xFF) << 56) |
-               (((long)b[offset + 1] & 0xFF) << 48) |
-               (((long)b[offset + 2] & 0xFF) << 40) |
-               (((long)b[offset + 3] & 0xFF) << 32) |
-               (((long)b[offset + 4] & 0xFF) << 24) |
-               ((b[offset + 5] & 0xFF) << 16) |
-               ((b[offset + 6] & 0xFF) << 8) |
-               (b[offset + 7] & 0xFF);
+        return (((long)b[offset] & 0xFF) << 56) 
+             | (((long)b[offset + 1] & 0xFF) << 48)
+             | (((long)b[offset + 2] & 0xFF) << 40)
+             | (((long)b[offset + 3] & 0xFF) << 32)
+             | (((long)b[offset + 4] & 0xFF) << 24)
+             | (((long)b[offset + 5] & 0xFF) << 16)
+             | (((long)b[offset + 6] & 0xFF) << 8)
+             | ((long)b[offset + 7] & 0xFF);
     }
 }
